Extract common testing into buildSrc
