+++ library/build.gradle.kts
@@ -1,4 +1,3 @@
-import org.gradle.api.tasks.testing.logging.TestExceptionFormat
 import org.jetbrains.dokka.gradle.DokkaTask
 
 plugins {
@@ -25,44 +24,8 @@ dependencies {
     "kaptTest"(Deps.Moshi.codeGen)
 }
 
-tasks.withType<Test> {
-    useJUnitPlatform()
-
-    testLogging {
-        showExceptions = true
-        showCauses = true
-        showStackTraces = true
-        exceptionFormat = TestExceptionFormat.FULL
-    }
-
-    reports.html.isEnabled = true
-    reports.junitXml.isEnabled = true
-}
-
-val check by tasks
-val jacocoTestReport by tasks
-
-tasks.withType<JacocoCoverageVerification> {
-    check.dependsOn(this)
-    dependsOn(jacocoTestReport)
-
-    violationRules {
-        rule {
-            element = "BUNDLE"
-            excludes = listOf("com.jacoco.dto.*")
-            limit("INSTRUCTION", 0.90)
-            limit("BRANCH", 0.5)
-        }
-    }
-}
-
-tasks.withType<JacocoReport> {
-    reports {
-        html.isEnabled = true
-        xml.isEnabled = true
-        csv.isEnabled = false
-    }
-}
+testWithJunit()
+coverWithJacoco()
 
 val dokka by tasks.getting(DokkaTask::class) {
     outputFormat = "html"
@@ -147,12 +110,5 @@ publishing {
                 }
             }
         }
-    }
-}
-
-fun JacocoViolationRule.limit(type: String, minValue: Double) {
-    this.limit {
-        counter = type
-        minimum = minValue.toBigDecimal()
     }
 }
